Update documentation
Forgot to document this. It's not as likely for the the user to output
multiple files under the "current" or "continuous" modes so the filename
for the initial file is kept clean while subsequent ones are tagged to
prevent overwrites. A-B mode is likely to be used to clip multiple
ranges so the first file is also tagged in order to maintain uniformity.

Also includes comment clarifications regarding internal variables as
well as information on runtime option changes.