Include AV1 in mp4 if compatible and format isn't detected as such
This allows AV1 in mp4 in cases where the format isn't a transport
stream and isn't detected as mp4, but the streams are still compatible
with it as an output format. Previously these streams would be written
to mkv. MPEG-4 as the container is preferred here as it has a more
precise timebase scale and thus more accurate frame timestamps, making
it more suitable for writing streamed content. AV1 also seems to be more
commonly associated with mp4 than mkv when the video host is queried for
metadata.

At the moment, outside of the format specifically being detected as MP4,
only AV1 + AAC and H.264 + AAC are supported for MP4 as they're the most
common combinations, but dump-cache has no problems writing other
combinations like AV1 + Opus to it most of the time (as long as there
are no codec_tag issues).

The codec combination detection is kept narrow for simplicity and in
order to avoid incompatible output format errors; MKV ultimately
supports a wider range of combinations and the compatibility issues will
soon be fixed in mpv anyway, see PR mpv-player/mpv#8859