Add force_extension option and remove transport stream container
Since the compatibility issues with the recorder/dump-cache have been
resolved in mpv-player/mpv@643c699 a new option has been added allowing
the user to force a preferred format and sidestep the automatic
detection. If using this option it is recommended that a highly flexible
container is used (e.g. Matroska). The format is specified as the
extension including the dot (e.g. force_extension=.mkv).

Transport streams have been removed from the automatic detection as they
can now be written to other containers if mpv is up to date. The script
now only tests for MP4 and WebM while falling back to MKV. Determining
common and standard formats is still useful if, for example, the user
plans on uploading the saved streams elsewhere as some sites only
accept/support specific formats.