Make the incrementing more logical
If you dumped a file using increments then changed the title or the
extension and dumped another file, the output file would have the same
increment as the previously dumped file.

There are 4 ways to fix this:

1. Reset file.inc at the beginning of increment_filename;
2. Reset file.inc if file.name is unique;
3. Reset file.inc anywhere file.title or file.ext change;
4. Increase the increment whenever a file is output.

Fix 1 isn't optimal if many files are being dumped as these files would
have to be looped through every time.

Fix 2 can get messy and requires testing file.name against a previous
filename which would need to be tracked.

Fix 3 is a bloated and inefficient solution considering all of the
places where these variables can change.

Fix 4 (this commit) is a simple way to make the increments follow a more
logical progression compared to the other three fixes (which would reset
the increments). Now the increments will increase for every file output
of the same video, regardless of the user set title or extension.

increment_filename has been slightly modified in order to account for
the increment increase at output so that an increment isn't skipped
under the "current" and "continuous" modes (as keeping the initial file
untagged is still preferred here).