Add force_title option
Works the same way as force_extension. This gives you the option to
set the filename title for all output files without having to send
runtime commands. Universally forcing the title might be desired in some
use cases especially since the script will prevent overwrites of files
with the same name.

Note that if you set force_title at runtime or revert the title to the
media-title the increments will be reset. Although this is a deviation
from the previous commit and inconsistent with the current general
behaviour, this does give the user more of a choice over how the
incrementation is implemented for different filename titles. The
increments will not reset if script-message commands are used to
override the title to a custom one instead.