Don't reset increment on runtime option changes
This case was overlooked:

1. set force_title
2. dump two files successively using a mode that isn't A-B
3. change any option at runtime
4. dump another file

In this case the first two files would be:
file.ext
file-2.ext

while the third file would be:
file-1.ext

This commit prevents this by making sure the increments are not reset
if the user options are updated at runtime.

Generating a separate set of increments for different titles is still
possible by setting force_title and then running:
script-message streamsave-title revert
twice in a row.