Option to add chapters at A-B loop points
The new range_marks option allows the script to set temporary chapters
at A-B loop points. If chapters already exist they are stored
and cleared whenever any A-B points are set. Once the A-B points are
cleared the original chapters are restored. The script also attempts to
take into account any chapters added after A-B mode is entered by adding
them to the initial chapter list.

This is meant to make clipping more intuitive by:

1. adding visible indication on the OSC of the cached range the script
will dump;

2. allowing you to easily move to the start & end points of this range;

3. providing a simple way to query the set range by right clicking the
chapter buttons on the OSC.

This option requires mpv to be built with mpv-player/mpv@6abb7e3 in
order for the chapter marks to immediately show on the OSC.

The option is disabled by default as A-B loops are used outside of cache
dumping and this option temporarily changes the chapter list.
Set range_marks=yes in streamsave.conf in order to enable it.