Use the same incremental tagging logic for all dump modes
The reason why the continuous and current modes had different labeling
rules has been reconsidered and the behaviour has been changed so that
all modes are now consistent and use the same rule for the primary file
if the label is set to increments.

This is preferable if you fully dump multiple files of the same stream
with the same title as the naming structure will be uniform. Having an
extra number in the filename for single file outputs is not a big deal.