Automatic stream saving
This adds an automated stream capture feature by allowing you to choose
whether the script automatically begins dumping the cache on stream load
and also by giving you the option to specify a stop time by setting a
cache time limit.

This is useful if you want to automate stream downloads or wish to
start writing the file immediately while watching streams concurrently.
The autoend feature lets you automatically limit the duration of the
output and thus also the file size.

Particularly, this commit is a prerequisite for the addition of a
workaround hack which facilitates saving some tricky streams where the
host changes. See the subsequent hostchange commit for more information.

Basically this is like passing --stream-record when opening a URL, but
with the script's benefits of having the file name and format set
automatically and the prevention of file overwrites. The main motivation
behind the addition of this feature is in providing a foundation for
dealing with streams which switch intermittently by allowing for
automatic cache writes.