Host change hack
This is a workaround for some DAI HLS streams served from .m3u8
which are characterized by mid stream host changes.
See mpv-player/mpv#5694 for more info.

Currently attempting to directly download these streams using
youtube-dl, yt-dlp, ffmpeg, or streamlink results in multiple issues
and broken files.

This currently experimental feature allows you to load the stream in mpv
and automatically save it piece by piece, making it possible to
reconstruct the whole stream.

How this works is that the script simply reloads the file whenever it
detects cache variations which are highly correlated with host changes.
As a result, mpv will write multiple files (one for every reload) which
you can then concatenate, mux, or simply play with mpv using
--merge-files or as a playlist.

The quit option is essentially a replacement for autoend (used for
direct streams), allowing you to set a one shot timer from script load
to when the player will exit. Because the autostart option is required
for this hostchange feature to work the player will continue writing
files until it closes.