Extract format values directly from observed properties
The format values are now passed on as arguments to the container
function as it's being called by changes in all three observed
properties. This way we can use the parameters instead of having to get
the properties again each time from within the function.

This was not done for the loop point variables as it's much better to
to get them directly for immediate use as several functions depend on
them and they need to execute in a linear, synchronous fashion.