Add piecewise option
This is mainly for saving long streams on slow systems or for people who
want to save a stream in parts automatically. On some slower machines
dumping a large cache can bog things down quite a bit until the writing
stops, so what this option does is allow you to dump the cache
periodically according to the time set in autoend.