container: get stream properties on demand (reverts 61325ae)
On consideration, the original logic was more sound. Since the output
format is only determined after the input format is detected and having
both audio and video tracks present is not required then it is necessary
to make sure we have the updated values before setting the container.

Since we now have automatic cache writing it is especially important to
make sure we don't attempt to write until all three stream properties
exist after load (esp. after reloading, on track and playlist changes)
and since the assigned stream variables persisted until the update then
it was possible for the previous values to be used in computing the
file extension.

Getting the properties directly guarantees immediate update and is
apparently faster than waiting for the change notification from the
observation wrapper. Under the latter scheme it would have been ideal to
call the file-format observer function last, as the latest values of the
audio and video streams are required for a proper determination of the
container.

However, on stream changes there was no guarantee that the file-format
would be updated last and so an incorrect extension could be assigned
temporarily until the new value was passed. Now, this happens fast
enough to where it doesn't matter under normal use, and even with
autostart enabled the first buffered cache time from the demuxer seems
to be approximated only after the decoders, formats, and protocol have
been determined.

That being said, it's good practice to simply make sure we're working
with the right properties even if it's redundant as the format
determination depends on all three properties and this could prevent
some edge cases in the future, especially under automatic cache writing
or if mpv's behaviour changes.