Reworking of write state and handling of same format track switches
Since we have redundant calls to the container handler we want to make
sure the write state isn't flipped unnecessarily. Basically, it should
now properly work as an autosave write request which is toggled in the
following scenarios:

- on automatic writing;
- if the cache time doesn't exist;
- at piece transitions (if piecewise is active);
- if the stream reloads or unloads;
- if the tracks switch;
- if seeking occurs.

For the last two there is an additional check which verifies that
there's been a cache reset of some sort.

The drawbacks to this implementation are namely that the write state is
flipped at quite a few places and that a variety of additional
properties must be observed once the cache is initially written.

A simpler, alternative approach could be to observe the cache state
itself and act on it, but that property changes very frequently and
contains quite a bit of data we don't actually need.

The current approach is much more specific, bordering on
micro-management, but should yield increased resiliency and accuracy
at the expense of added complexity.

In short, we want to auto-write files only when necessary while covering
all of the cases in which we should have a pending write request.