cache_write: use mp.command_native_async instead
This might be a better way to do it. Now, I'm not sure if there are any
performance benefits and I haven't looked at the mpv source code, but
there are differences in the timing characteristics between mp.commandv
with the async prefix and mp.command_native_async for dump-cache.

It seems as though mp.commandv, even with the async prefix, still blocks
the caller until at least the writing has been initialized, which is why
it was possible to verify the existence of the file immediately after
running the command, as if the program waits for a reply that the file
writing has started before continuing.

In contrast, mp.command_native_async is completely asynchronous.
Using this command instead it becomes possible to provide more detailed
diagnostics on callback through the command-reply event, which happens
once the writing has been completed or there's an error.

This also allows us to use a proper stop function. Previously, writing
was stopped on demand by attempting to write another file using an empty
string as the file name. Because mpv's recorder can only write one file
at a time, this would stop any previous writing, but would not write
another file. This, however, requires the demuxer to be open which
wasn't always the case on resets. We also had to rely on the write
request state to determine if there's a file being written.

Anyways, if we're going to have automatic cache writing it's a good
idea to work with a cache write queue and run completely asynchronous
write commands.