Revision of the incrementation system
So, this needed a revisit for two reasons:

1. Now that the cache writing is completely asynchronous advancing the
increment at file initialization is a little bit tricky. Not doing so
results in the same incremental tag being used if the format changes.

2. Automatic cache writing under hostchange makes use of reloads which
before this commit would reset the file increments. This is inefficient
especially if there's a large number of host changes as then all of the
existing files have to be looped through to set the next file name.

Making use of oldtitle and oldext by setting them at the change
handlers doesn't work universally as some videos will set a different
media title temporarily on load/unload. So the simple solution is to
reset the index only when the file name changes.