Prevent overwrites using range label
Most of the time overwrites which happen using the range label usually
dump the same data, but this is not necessarily the case if a recurring
stream consistently uses the same title in which case file names can
clash for the same output range. So append increments if the file exists
in order to be safe.