Add option to track HLS packet drops
This option adds chapters to positions where packet loss occurs for HLS
streams. This gives you a visual sense for where the video is choppy due
to low bandwidth or during times of network congestion and makes it easy
to preserve continuity and trim around skips which usually cause invalid
timestamps. There's probably a better way to keep track of this, but I
couldn't find the relevant property so for now we're just going to parse
log messages.