Register the chapter_points callback only on the initial load
Either reloading the stream or loading another file at runtime would
register another set of callback mechanisms associated with changes to
the set of loop points which would run the chapter_points procedure
multiple times redundantly if the points were reset. Preventing this is
especially important under automation since the hostchange feature makes
use out of reloading the stream and the piecewise option automatically
sets loop points and so the number of useless calls to chapter_points
could be significant.