Rework how maintenance is conducted on the segments table
The previous method of constructing and annihilating the table was too
convoluted and was prone to error due the extra layer of nested
separation. Particularly, the 'continuous' clause at the initialization
of the cache_write process is there to ensure that any subsequent
request to write a file when dumping is continuous is respected (as
writing is strict under chapter modes and so everything else gets
queued), but it also provided a bridge into the routine which could end
up attempting to service an empty table if the previous file had not
stopped writing.

For some reason I found it useful to have a distinction between having
completed the writing and having processed the last segment for writing,
and under most cases this was covered, but there were some possible edge
cases where the last segment gets sent out for writing and the callback
hasn't returned yet where the cache_write process could be entered with
the segment list in limbo; although this wasn't verified if for whatever
reason the duration goes missing during the construction of these
segments then writing falls back into a continuous mode which could lead
to the aforementioned problems.

So this is really a long-winded way of saying don't destroy the
segments table as its existence is critical for much of the write
preparation. There were no obvious issues with the previous
implementation, but the distinction between writing the last piece and
having written it (empty table vs. no table) isn't necessary as it's not
used anywhere and could lead to errors.