Use match instead of find, tighten restrictions on the input time format
So since we don't need the position anywhere we really should have been
using match instead. Although I suppose using find for the first case
could be somewhat easier to read as checking the existence of the
substring against the index is slightly more intuitive than the implicit
case where comparison with any missing captures means the string wasn't
matched; this isn't a great justification however, and it could easily
be argued that the converse is easier to follow so we're just going to
do it the proper way. Also the match is more restrictive now in order to
reduce any possible unexpected results and confusion on the user side,
since we're validating the input format anyway.