get_chapters: return the update state
Okay so thinking through the previous chapters commit made the following
pretty clear - if the loop points were unset the chapter-list property
would be set regardless if the stored chapters matched the current
chapters which is why if a new stream had no chapters present its
property would be superseded by the previous list and would have its
empty chapter list overwritten. This is because chapter_points is a
listener for both the loop observables as well as stream load events.

Although this was fixed (by also capturing empty chapter lists) it's
obvious that setting this property when the two lists match is
unnecessary especially since it is unclear as to what happens exactly on
natively setting the chapter-list array. Work on the packet tracking has
led to the suspicion that the player can choose to ignore a request to
set the chapters (more on that later in a subsequent commit) and it's
unclear as to why that is, so we should avoid setting it when we don't
need to. Anyway, this also simplifies the decision to set the chapters
at fragment_chapters.

In essence, the last chapters commit (7f04124) solved the problem of the
property being set when the chapter lists were mismatched with the push
not representing the latest list while this commit solves the problem of
the property being set while the two are in fact equal.