cache_write: various optimizations for the new chapter write modes
This brings execution of write_chapter entirely under cache_write,
tightens the restrictions on the active file queue under a live
segment_list, and suppresses the initial OSD message under the segments
mode. This should address and prevent various edge cases under quick
writes by making sure the writes are executed in a controlled fashion
as a security measure.

As for the messages, most of the OSD spam was already suppressed except
for the initial message, but this message was just as important as it
also signalled on the completion of the initial chapter and this could
lead to a miscomprehension especially if a large cache is being dumped
(which can hold up some threads even though the implementation is as
asynchronous as possible); we only want to communicate that the cache
dumping is complete on the OSD after the last chapter has finished
writing.