track_packets: avoid comparing floats and using floating indices
So that's almost always a terrible idea, and in our case indexing with
floats would sometimes cause the script to crash (about 1% of the time)
due to floating point errors apparently. An easy way around this is to
simply count each time the event fires and track each pending chapter
by stamping the title with the event id. Not as clean I guess, but
should be much safer.