hostchange overhaul - exploit track switches instead of cache variations
Instead of adding additional conditions each time a false positive or
false negative is encountered (e.g. 3e516ca and 81b8856 etc.) it's best
to just rework the entire feature using a more reliable property
(in this case, track-list).

Although this new approach wasn't tested with a wide variety of live
streams, it should generally prove to be more robust; that being said,
the original implementation may have its advantages if the inserts do
not trigger demuxer events.

This should increase the reliability of the insert detection, reduce
errant reloads, and dispatch with unnecessary complexity. The automatic
handler was doing quite a bit of heavy lifting which was not ideal as it
was overseeing all functionality behind autostart, autoend, piecewise,
and hostchange, but it made sense to keep everything under the same
umbrella as long as hostchange required the utilization of cache
timestamps.

As long as stream switches trigger track-list updates without id
changes, however, this should allow us to detach the detection and
reload mechanisms and make them practically independent from the rest of
the automation features.