Choose WebM over MKV if AV1 is present with WebM compatible audio
Although apparently not officially part of the standard, AV1 is now
widely supported inside of WebM and has become part of the de facto
specification.

This currently shouldn't make a difference for most streams as usually
when AV1 video is present the format is detected as MP4, but if that
header isn't read as MP4 then WebM will now be preferred for output over
MKV as long as the stream contains Opus or Vorbis audio. This is ideal
for compatibility with browsers, etc.