Refine the list of allowed codecs for MP4 output to a standard set
This now includes the standard formats of H.265, MP3, and FLAC, but
restricts the set to only the most compatible combinations. Although the
MP4 container does support codecs such as VP9 and Opus, and mixing and
matching audio/video streams with yt-dlp is quite common, mpv's recorder
is prone to errors when writing the cache into that container for some
of these codecs.

For example, VP8 will refuse to write to MP4, VP9 will output a broken
file, and including Opus will cause a slight a/v desync likely due to
some strange timestamp issue. MKV, on the other hand, doesn't suffer
from these issues. MP4 is still preferred over MKV, however, when the
formats are safe in order to maintain web standard compatibility.

Watching the file format has been done away with as well as string
matching on it (as that would still cause MP4 output under the
issue-prone combinations) since it's no longer needed and now only the
codecs are used to determine the output format; the property is still
extracted as it's useful for checking initialization.