Call observe_tracks under force_extension whenever the formats change
I guess I must have just overlooked this because it seems to be missing
and yet necessary since it's called whenever the extension isn't forced.

There should be a way to simplify the logic here, but the automation
code with hostchange is a bit finicky and I don't want to break it as
debugging it is pretty involved so just adding this line should be safe.