Reset the file extension when the file format is missing
The file format is missing before the stream is fully loaded on the
initial load, on unload, on reload and when loading another stream.

Resetting the extension here is mostly a safety mechanism as during this
time the demuxer is not open so file writing cannot be initiated and
under our automatic dumping mode commencement waits on the existence of
the cache.

Either way, the extension shouldn't be carried over and this prevents
attempts to write the cache until we have a valid format and the
stream is properly loaded.