Add the ability to set force_[title|extension] through script-message
So I've decided against removing the updating of options using
script-opts at runtime because it is in fact useful and serves a
specific purpose. Even though there are two different ways of doing
practically the same thing (and with different parameter names which may
or may not be advantageous) there's something you could easily do
through script-opts which is not exactly possible to do using
script-message: resetting / clearing non-default options.

If you load a stream passing specific options using --script-opts it's
easy to clear them and reset the options to what's specified in the
script's config file using `set` or `change-list`. To accomplish this
using script-message you have to send the commands individually for each
option you changed at the beginning, there's no way to simply reset.
This is because script-opts is applied on top of the config file and is
a list option so it makes this very simple as long as runtime updates
are supported.

So even though script-message is a bit more robust and convenient,
supports status messages and extra arguments, script-opts becomes useful
if you're modifying a wide swath of options at the start and want the
ability to revert to your own specific defaults with a single command at
runtime. Maintaining them isn't much of an extra burden anyway.

As for this specific commit, these extensions to the title and format
script-message override functions contain what was missing to do away
with script-opts outside of what I've delineated in the preface above.

Even though it's possible to change these through script-opts doing this
using script-message is as said more convenient. So if you open a stream
and want to change the format and set it for all subsequent streams
loaded you can now do (for example):

script-message streamsave-extension .mkv force

and it will set force_extension without having to go through options.