Omit the media title under the timestamp label
The original format wasn't of much use and had only been implemented in
order to prevent overwrites, but it hasn't been used for that since the
earliest versions of the script. Appending the timestamp makes the file
name unnecessarily long and I doubt anyone was really using this label.

This way the file name will simply be the timestamp itself, which does
make it somewhat useful.

Also, set_name has been modified such that it can be used universally in
this cases (as is also the case using the chapter label).