chapter modes: refuse to write and throw an error if not fully cached
Attempting to write out uncached chapters could result in a write loop
which would apparently continue to write broken files; although usually
this would write empty files for the given number of chapters (which
itself is not so great) if the former error is triggered hundreds of
write requests would be initiated which is particularly nasty. There's a
chance this might have something to do with the queue system in place
under write failures, and this should definitely be investigated, but
for now this should be prevented by practically throwing an exception if
the chapters are not fully cached.

Also, instead of falling back to continuous writing if no chapters are
present under the segments mode the script will now throw an error.

The cache checking has been made robust as there's no guarantee that the
duration property and the end cache timestamp match exactly; the cache
timestamp could be more accurate than the duration, and the duration
could be different from the eof timestamp as well if the stream's start
timestamp is not zero. Although we do use a time of zero for the zeroth
chapter, this should generally be fine. The end timestamps for the last
chapter will now be set from the seekable ranges maximum timestamp,
either using get_seekable_cache for the segments mode, or essentially
doing the same thing within the cache_check loop for the single chapter
mode.

Quite a bit of complexity has been added here in order to be safe, so
there's a good chance that I'll remove the chapter modes for the lite
version of the script; having them included as part of the mode-switch
cycle is also kind of bothersome if you don't use these modes anyway.