title_override: set the output_label to increment if forced
Since both the chapter and timestamp labels disregard the file title and
effectively set the file name the question arises as to whether
force_title should take precedence. On the one hand, it's called
force_title so there's some implicit understanding that this should be
what's used regardless of label. On the other, it's not called
force_name and as it currently stands it does force the title for the
labels which use it. Changing the current behaviour if running
force_title from the outset would make it more difficult to change the
label at runtime if a chapter or timestamp filename is desired.

So as a compromise if the title is forced through script-message then
the label will be switched over to increment in order to match.