Refine the cache checking so the end chapter is written continuously
This avoids having to set the end point timestamp from the cached
seekable ranges and allows the modes to be used for live streams
(where eof is false until the stream ends). The latter will be made
useful with the next commit which will open up the modes for use with
range_mark chapters.