Relax the restriction on chapter list extraction for the chapter modes
This gets the current chapters directly which means that the A-B loop
point chapters are now viable for writing under the chapter modes.
Previously calling get_chapters would only use the master chapter_list
which would disregard the loop point chapters if set under range_marks.

Script set chapters are written if using track_packets so this brings
things in line with this convention.

This is useful if we want to divide a live stream in such as a way as to
cut a desired clip in a given interval while also dumping the entire
stream split into two or three parts. This is now accomplished by
setting the desired loop points (either one or both will work) and then
dumping under the segments mode; this spits out the clipped range and
the rest of the parts avoiding redundancy (e.g. having to cut in A-B
mode and then continuous mode).