Fix and simplify the cache checking for the chapter modes
The first two approaches (fba819e and 180b1c4) were effectively broken
as they didn't cover all of the cases. This should now work as it uses
the end chapter if present and enables continuous writing for it by
checking that the cache end timestamp for the last range exceeds the
end chapter's start timestamp. This allows us to eliminate the branch
with the end chapter loop while ensuring that chapters aren't sent for
writing unless the cache is suitably prepared.