Set the zeroth chapter start time to the seekable cache start time
For HLS streams using --rebase-start-time=yes (the default) for some
reason the timestamp will not be exactly zero and will instead vary by
a few milliseconds which meant that the chapter write modes would fail
even though the stream or chapter was essentially fully cached.

There does not seem to be a property matching this start time
(demuxer-start-time returns the random timestamp prior to rebasing)
and I'm unsure as to whether this is a bug (the manual does say it's
supposed to be zero) so I'll have to do some more testing in order to
figure out why there's a slight mismatch (it's usually on the order of a
frame interval or so).

In any case, this commit defines the complement to get_seekable_cache
(which returns the very last cache timestamp and is more accurate than
demuxer-cache-time) so that we can use the real stream start time for
the zeroth chapter if present.

The return order for get_ranges has been swapped to better reflect
the name and since for get_cache_start we don't need the cache_state.