Properly implement threshold cache checking for the chapter modes
After some testing it's apparent that the stream's initial seekable
cache timestamp can indeed be affected by the aforementioned offset
and that this time can differ from the initial playback-time possibly
due to rounding errors. So,

- use the length of segments rather than chapter_list
- for the single chapter mode set the zeroth chapter end timestamp based
  on this threshold
- for the multiple chapter mode only cache check if there's more than
  1 chapter to write

for the latter the threshold is still used in deciding whether to insert
the zeroth chapter, and if it isn't inserted and k is 1 then up_to_end
(whether the last chapter's start time exceeds the last seekable range's
start time) doesn't need to be checked since in this case the initial
chapter's start time is effectively being treated as matching the stream
cache's initial timestamp.

Fixes 7d64f71