Adjust initial chapter acc. threshold, reintroduce get_cache_start
This has proved to be trickier than I initially thought. There's quite a
bit of complexity now meant to cover possible edge cases and while I'm
sure there has to be a simpler way to accomplish the chapter mode cache
checking this works and will have to do for now.

This commit reverts most of the last commit (4cd270c) while keeping the
n --> k chapter_list to segments length exchange, and cherry-picks
get_cache_start (with some modifications) from 3a1aebc so essentially
this also reverts the revert at 7d64f71.

The solution is ultimately straightforward enough in idea although not
in its implementation I suppose. Basically if the initial chapter's
start timestamp is offset from zero by less than 100 milliseconds then
we treat that as matching the stream's start and adjust the chapter's
timestamp accordingly using the seekable cache's origin. This is easy to
do given that we're not modifying the master chapter_list, but instead a
local version of the current chapters extracted for writing.

This makes the rest of the comparisons less complicated and prevents
writing unless properly cached (avoiding broken files and errors etc.)
while also allowing for writes when available.

Completes fba819e