Trim the script down to a lite version
Since there's a significant amount of complexity contained within the
main script and many of the extra features are most likely not used by
the majority of the userbase I've decided it's a good idea to provide a
base script with the core essentials and a few auxiliary features
related to them. This should also help in debugging the main script.

What has been stripped includes:
- All of the automation features;
- The packet tracking feature;
- script-message commands for most options.

Runtime updates to options are still possible through script-opts.

The range_marks option here is enabled by default. Since this option
manipulates the chapter list whenever loop points are set it is disabled
by default in the main script, but since this version is not at the
master branch and this option is part of the base functionality I've
decided to enable it here.

This version will depend on bug reports and feedback more than the
master version as I personally use the latter. However, given that the
lite version is for the most part a subset of the master version any
bugs present here should be reflected at master.