save_directory: accept paths with a trailing slash
This allows for a bit more flexibility in specifying the path at the
user end at the expense of added complexity. Previously it was necessary
to remember to elide the slash and there was no validation performed in
the case it wasn't. Now it can be included or not, and if it isn't then
it's appended.

append_slash is called twice in order to cover expanded paths.
Without the initial call the path shortcuts by themselves without the
trailing slash will not be expanded (e.g. save_directory=~~), and
without the secondary call if these shortcuts are expanded then they
will be missing the slash.